+++ packages/typespec-rust/test/spector/azure/resource-manager/method-subscription-id/tests/method_subscription_id_two_subscription_resources_method_level_subscription_resource2_operations_client_test.rs
@@ -21,7 +21,7 @@ async fn subscription_resource2_get() {
         .await
         .unwrap();
 
-    let resource: SubscriptionResource2 = resp.into_body().unwrap();
+    let resource: SubscriptionResource2 = resp.into_model().unwrap();
     let expected_resource = get_valid_subscription_resource2();
 
     assert_eq!(expected_resource.id, resource.id);
@@ -51,7 +51,7 @@ async fn subscription_resource2_put() {
         .await
         .unwrap();
 
-    let created_resource: SubscriptionResource2 = resp.into_body().unwrap();
+    let created_resource: SubscriptionResource2 = resp.into_model().unwrap();
 
     let expected_props = resource.properties.unwrap();
     let created_props = created_resource.properties.unwrap();