+++ packages/typespec-rust/test/spector/payload/xml/tests/xml_model_with_attributes_value_client_test.rs
@@ -12,7 +12,7 @@ async fn get() {
         .get(None)
         .await
         .unwrap();
-    let value: ModelWithAttributes = resp.into_body().unwrap();
+    let value: ModelWithAttributes = resp.into_model().unwrap();
     assert_eq!(value.enabled, Some(true));
     assert_eq!(value.id1, Some(123));
     assert_eq!(value.id2, Some("foo".to_string()));