+++ packages/typespec-rust/test/spector/payload/xml/tests/xml_model_with_unwrapped_array_value_client_test.rs
@@ -12,7 +12,7 @@ async fn get() {
         .get(None)
         .await
         .unwrap();
-    let value: ModelWithUnwrappedArray = resp.into_body().unwrap();
+    let value: ModelWithUnwrappedArray = resp.into_model().unwrap();
     assert_eq!(
         value.colors,
         Some(vec![