+++ packages/typespec-rust/test/spector/payload/xml/tests/xml_simple_model_value_client_test.rs
@@ -12,7 +12,7 @@ async fn get() {
         .get(None)
         .await
         .unwrap();
-    let value: SimpleModel = resp.into_body().unwrap();
+    let value: SimpleModel = resp.into_model().unwrap();
     assert_eq!(value.age, Some(123));
     assert_eq!(value.name, Some("foo".to_string()));
 }