+++ packages/typespec-rust/test/spector/type/property/optionality/tests/optional_required_and_optional_client.rs
@@ -12,7 +12,7 @@ async fn get_all() {
         .get_all(None)
         .await
         .unwrap()
-        .into_body()
+        .into_model()
         .unwrap();
     // According to mockapi.ts, the all endpoint returns
     // { optionalProperty: "hello", requiredProperty: 42 }
@@ -29,7 +29,7 @@ async fn get_required_only() {
         .get_required_only(None)
         .await
         .unwrap()
-        .into_body()
+        .into_model()
         .unwrap();
     // According to mockapi.ts, the required-only endpoint returns { requiredProperty: 42 }
     assert_eq!(resp.optional_property, None);