+++ packages/typespec-rust/test/spector/type/property/value-types/tests/value_types_boolean_client.rs
@@ -12,7 +12,7 @@ async fn get() {
         .get(None)
         .await
         .unwrap()
-        .into_body()
+        .into_model()
         .unwrap();
     assert_eq!(resp.property, Some(true));
 }