+++ packages/typespec-rust/test/spector/type/property/value-types/tests/value_types_collections_int_client.rs
@@ -12,7 +12,7 @@ async fn get() {
         .get(None)
         .await
         .unwrap()
-        .into_body()
+        .into_model()
         .unwrap();
     assert!(resp.property.is_some());
     if let Some(collection) = resp.property {