+++ packages/typespec-rust/test/spector/type/property/value-types/tests/value_types_never_client.rs
@@ -12,7 +12,7 @@ async fn get() {
         .get(None)
         .await
         .unwrap()
-        .into_body()
+        .into_model()
         .unwrap();
     // The NeverProperty struct should have no properties
 }